Factor out renderer
